Port PTB, the XML container and KEBA
KEBA is a SAFE XML container carrying a hundred OCMF documents that group
into one charging session of 190 readings -- by far the largest fixture,
and it passed unchanged except for the session id.

That exception was my own doing. In the chargeIT step I had added
FloatParseHandling.Decimal alongside the date fix, reasoning that a meter
reading should not become a binary float. It looks like the safer choice
and is not: a decimal remembers that a meter wrote "0.00" rather than "0",
canonical JSON renders numbers the ECMAScript way, and the OCMF session
identity is a hash over that canonical form. Preserving the trailing zeros
gives the same charging session a different identity from the one every
other implementation computes. The precision a double costs is irrelevant
here -- meter readings stay far inside the fifteen significant digits a
double round-trips exactly -- while the identity is not negotiable. The
setting is gone and the reasoning is written down where it was wrong.

PTB is a small envelope around two OCMF documents plus the one thing OCMF
cannot say: where the charging station stands. Its schema is checked
strictly and every violation is reported at once rather than the first --
the place is precisely what somebody would have to falsify to bill a driver
for a charging session at a station they never visited, and a reader that
gave up after the version number would still be busy with that while the
address was wrong.

Which surfaced the second finding: OCMF was discarding everything the
container knew. The address, the geographical location, the description,
the firmware -- all dropped, the charging station rebuilt from
identifications alone. No golden file prints an address, so nothing caught
it. That is the whole reason a container exists, and an EV driver shown
only a meter serial number had been told less than the file contained. The
SAFE containers were losing it too, not just PTB.

The XMLContainer is ported faithfully and stops where ChargyCore.TS stops:
it reads the container, checks that one public key, one signature method
and one encoding hold throughout, and then reports that it cannot turn the
values into a charging session. Upstream leaves that conversion as a ToDo,
and inventing one here would mean claiming to verify something the
reference implementation does not. A test pins it, so the day upstream
finishes the conversion it fails and says so.

Finally, the two PTBContainer/*.expected.txt files turn out to be orphans
in both repositories: their report format ("format: ptb",
"energyDifferenceWh") is produced by no code in ChargyCore.TS, and the
three tests that would use their fixtures are test.skip -- the fixtures are
hand-made demo data that OCMF itself rejects. The reachable golden-file
count is 23, not 25. A test pins that the PTB envelope accepts them and
OCMF is what turns them down, because the fixtures are checked in, look
usable, and are not.

445 tests, 0 warnings.

+++ ChargyCore/ChargyLib.cs
@@ -521,6 +521,17 @@ namespace cloud.charging.open.chargy
         /// their timestamps as text, several formats keep the meter's own UTC
         /// offset in them, and a re-rendered timestamp is a different timestamp.
         /// So the reader is told to leave strings alone.
+        ///
+        /// Numbers are deliberately left as doubles, which is Newtonsoft's default.
+        /// Reading them as decimals looks like the safer choice and is not: a
+        /// decimal remembers that a meter wrote "0.00" rather than "0", canonical
+        /// JSON renders numbers the ECMAScript way, and the OCMF session identity
+        /// is a hash over that canonical form. Preserving the trailing zeros
+        /// therefore gives the same charging session a different identity from the
+        /// one every other implementation computes. The precision a double costs
+        /// is irrelevant here — meter readings stay far inside the fifteen
+        /// significant digits a double round-trips exactly — while the identity is
+        /// not negotiable.
         /// </summary>
         /// <param name="Text">A JSON document.</param>
         /// <exception cref="Newtonsoft.Json.JsonReaderException">When the text is not valid JSON.</exception>
@@ -529,7 +540,7 @@ namespace cloud.charging.open.chargy
 
             using var reader = new JsonTextReader(new StringReader(Text)) {
                                    DateParseHandling  = DateParseHandling.None,
-                                   FloatParseHandling = FloatParseHandling.Decimal
+                                   FloatParseHandling = FloatParseHandling.Double
                                };
 
             var json = JObject.Load(reader);
